+++ src/pages/api/payments/create-billing-portal.ts
@@ -2,8 +2,8 @@ import { NextApiRequest, NextApiResponse } from 'next';
 
 import { stripe } from '@/libs/stripe';
 import { supabaseServerClient } from '@/libs/supabase/server';
-import StripeService from '@/services/stripeService';
-import SupabaseService from '@/services/supabaseService';
+import StripeService from '@/services/stripe';
+import SupabaseService from '@/services/supabase';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     if (req.method !== 'POST') {
