+++ src/pages/api/webhooks.ts
@@ -2,8 +2,8 @@ import { NextApiRequest, NextApiResponse } from 'next';
 
 import { stripe } from '@/libs/stripe';
 import { supabaseServerClient } from '@/libs/supabase/server';
-import StripeService from '@/services/stripeService';
-import SupabaseService from '@/services/supabaseService';
+import StripeService from '@/services/stripe';
+import SupabaseService from '@/services/supabase';
 
 const endpointSecret = process.env.STRIPE_WEBHOOK_SECRET;
